DJ_A2_05 : Moved numberFormatter into utilities
