+++ src/containers/Sales/Sales.helper.ts
@@ -1,14 +1,4 @@
-/**  Formats and abbreviates large numbers (e.g., 1,200 -> 1.2K , 1,000000 -> 1M). */
-export function numberFormatter(value: number) {
-    if (value >= 1e6) {
-        return `${value / 1e5}M`;
-    }
-    if (value >= 1e3) {
-        return `${value / 1e3}K`;
-    }
-
-    return `${value}`;
-}
+import { numberFormatter } from '@utilities';
 
 /**
  * Formats the name and value according to custom logic or display preferences.
