+++ src/utilities/index.ts
@@ -1,1 +1,2 @@
 export { debounce } from './debouce';
+export { numberFormatter } from './formatter';
